+++ x2gbfs/providers/__init__.py
@@ -7,4 +7,3 @@ from .free2move import Free2moveAPI, Free2moveProvider
 from .lastenvelo_fr import LastenVeloFreiburgProvider
 from .moqo import MoqoProvider
 from .noi import NoiProvider
-from .voi_raumobil import VoiRaumobil